Add dependency Injection for Azure Search
This commit is to add the ability to inject your own Azure Search Client and Azure Search Index Client into this class. This allows the consumer of the Vanna library to chose how they would like to instantiate their clients, say with Managed Identity instead of being forced to use an API Key.